add get user data API
