+++ server/routes/userRoutes.js
@@ -1,5 +1,6 @@
 const express = require('express')
-const { registerController, loginController, checkDuplicateController, resetPasswordController, checkUserResetPasswordController, requireSignIn, registerDoctorController } = require('../controllers/userController')
+const { registerController, loginController, checkDuplicateController, resetPasswordController, checkUserResetPasswordController, requireSignIn, registerDoctorController } = require('../controllers/authController')
+const { getUserByUsername } = require('../controllers/userController')
 
 //router object
 const router = express.Router()
@@ -15,6 +16,10 @@ router.post('/login', loginController)
 // FORGET PASSWORD 
 router.post('/checkUserResetPassword', checkUserResetPasswordController)
 router.put('/resetPassword' /*,requireSignIn*/ , resetPasswordController)
+
+//GET USER data by username
+//GET /api/user/{username}
+router.get("/user/:username", getUserByUsername);
 
 //export
 module.exports = router
